BoxPerimeterRayCaster: Extract out oriented bounds into a separate file
