+++ Assets/PenguinQuest/Code/Controllers/AlwaysOnComponents/OrientedBounds.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0a40445b859c58345b2cecac15e351b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
